[refactor] apis와 interface 파일 일치화
- 기존 goods.ts의 타입들을 apis에 맞게 분리
