+++ src/interface/ingredients.ts
@@ -1,0 +1,6 @@
+export type Ingredient = {
+  id: number;
+  name: string;
+  imageUrl: string;
+  flavorIdList: number[];
+};
