+++ src/interface/recipes.ts
@@ -1,0 +1,11 @@
+export type Flavor = {
+  id: number;
+  flavorName: string;
+  imageUrl: string;
+};
+
+export type Recipe = {
+  id: number;
+  recipeName: string;
+  flavors: Flavor[];
+};
