+++ src/interface/sizes.ts
@@ -1,0 +1,6 @@
+export type Size = {
+  id: number;
+  name: string;
+  imageUrl: string;
+  value: number;
+};
